Restructure scheduler loop and basic version logic
This switches the scheduler to looping over the
activities to find new jobs.  This logic is a bit
easier to follow.

This also adds some logic to avoid submitting new
jobs for versions that are only different at the patch
level.

This also includes several fixes to the other classes.