Switch to using data models (#2)
* Refactor nmdcapi

- Adds minter method
- Handles token refresh in a better way

* Add workflow class

* Refactor initialization for wfutils

* Refactor sched

- Moves minting to nmdcap
- Switches to using the workflow class

* Remove refresh token calls

* Sync up names and test data

* Switch to requests_mock

* Cleanup mocks

* Cleanup mocks

* Add classes for key objects

This adds object class for activities, workflows, and data objects

* Refactor scheduler to use object classes

This moves the scheduler to use object classes instead
of dictionaries to represent and handle much of the core
scheduling.

The previous method is still in the code.  This will be
dropped after testing.

* Remove old version and dead code

Remove the old job finder approach and some minor flake8 cleanup

* Cleanup and fix sequencing git url

These cleans up some comments and adds
the git url for the sequencing activity.