Fix logic for in range activities
This addresses issue #211.  The fix is to load all
activities with the same git_url and then check
if the activity version is within range of version
of the workflow.  Tests were added to check the
behavior.

Some other minor cleanup is included such as
suppressing warnings for the same thing repeatedly.