Fix filtering for omics records
This fixes a bug when using an allow list.  Previously the scheduler
would construct the query filter and pass that to the load_activities
function.  Now it passes the allow list as a set and lets that function
create the query.  This is cleaner and the internal function can now
create the correct query depending on the collection type.