fix restore_from_state
The restore_from_state function was calling get_workflow_jobs_from_state which was only returning NEW jobs - i.e not already in the cache.  restore_from_state was setting value of self.job_cache to this return value, instead of extending the lob_cache with the resutls.  This resulted in overwriting the cache after the first cycle beat